Add delete, and edit functionalty to table

+++ scripts/init.sql
@@ -4,7 +4,7 @@ CREATE TABLE employees (
     id UUID NOT NULL DEFAULT uuid_generate_v4(),
     first_name VARCHAR(128) NOT NULL,
     last_name VARCHAR (128) NOT NULL,
-    salary INTEGER NOT NULL,
+    salary NUMERIC NOT NULL,
     created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
     updated_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
     CONSTRAINT pkey_item PRIMARY KEY(id)
